error render pero algo muestra

+++ src/main/java/com/grupo5/institutoEducativo/entity/Categoria.java
@@ -42,7 +42,7 @@ public class Categoria implements Serializable {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Basic(optional = false)
     @Column(name = "id")
-    private Integer id;
+    private Long id;
     @Basic(optional = false)
     @Column(name = "nombre")
     private String nombre;
@@ -55,21 +55,21 @@ public class Categoria implements Serializable {
     public Categoria() {
     }
 
-    public Categoria(Integer id) {
+    public Categoria(Long id) {
         this.id = id;
     }
 
-    public Categoria(Integer id, String nombre, float importePorHora) {
+    public Categoria(Long id, String nombre, float importePorHora) {
         this.id = id;
         this.nombre = nombre;
         this.importePorHora = importePorHora;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
